Correct new overlay code ignoring alpha in transitions
The new overlayer was inadvertantly storing the passed in alpha value
as an int; this bug showed up in the single-track overlapping
conversions, such as dissolves.  Corrected alpha type to float.

+++ cinelerra/overlayframe.h
@@ -108,7 +108,7 @@ class DirectEngine : public LoadServer{
   int out_x2;
   int out_y1;
   int out_y2;
-  int alpha;
+  float alpha;
   int mode;
 };
 
@@ -133,7 +133,7 @@ class NNEngine : public LoadServer{
   int out_x2i;
   float out_y1;
   float out_y2;
-  int alpha;
+  float alpha;
   int mode;
 
   int *in_lookup_x;
@@ -162,7 +162,7 @@ class SampleEngine : public LoadServer{
   float out1;
   float out2;
 
-  int alpha;
+  float alpha;
   int mode;
 
   int *lookup_sx0;
